Spawner Functionality ++
Spawner received a vast overhaul:
- ability to spawn different types of enemies
- ability to control exact sequence of enemies
- automatically controls interval between enemy spawns for a more cinematic gaming experience

PLEASE NOTE VERY IMPORTANT:
This version includes an experimental prototype of the wave system which is disabled by default.
To enable it please find the related variable that activates it and select it (making it true) in the inspector. without this selected the gameplay may result in a very imbalanced and frustrating experience. thus the developer STRONGLY recommends activating the aforementioned variable, if you are having trouble locating it please follow the instructions enclosed within the Spawner.cs file.